Change structure to make imports work correctly.
This commit overhauls the whole file structure of the project.
All sorts are moved from root_folder to basic_sorting/sorts/, and
all tests are moved from root_folder to basic_sorting/tests/.

In addition to that, I have added setup.py. Setup now consists of running
"pip install -e ." in the root folder, in order to create an editable
package, to use files from sorts/ in tests/

Thus, I also added .gitignore, to ignore .egg-info files that are
generated in the process.